Updated Shader Class.
A couple of changes were made to the shader class.

It seemed desirable to be able to create a shader object and then load
the shaders later. This is why a default constructor was added.

A new method GetUniformLocation() was added to make it easy to get the
locations of uniforms variables for the current shader. This method also
caches lookups so it doesn't have to ask the driver every time, which
can be expensive.

Changed create_program -> LoadShaders()

Added more comments and fixed class variable names so they start with
'm_'.