Camera is more or less done.
All of the basic features of the camera work now. In the future some
more things may be added, like infinite perspective. It was changed to
use a Quaternion to store rotations because this camera allows rolling
unlike the default cameras. Also the Pitch/Yaw act like a FPS. The Yaw
method was modified to use a special UP vector that's only changed
during rolling. Otherwise if you kept using pitch then yaw the scene
would slowly roll over, which is not desirable.