I lied the camera is still wrong.
So right now it's been fixed to be a 6DOF camera, but it has the
undesirable effect of still rolling when pitching/yawing. This is (I
believe) correct behavior, but not expected behavior. A work around
still has to be devised, maybe some corrective unrolling if there was
now actual roll angle present.